Add cache capabilities to get_predictions

+++ get_predictions.py
@@ -170,7 +170,14 @@ def check_for_unknown_painting(num_matching_descriptors, unknown_painting_thresh
     return False
 
 
-def get_predictions(query_dir, bbdd_dir, method, matching_method, matching_params=[], unknown_painting_threshold=1):
+def get_predictions(query_dir, bbdd_dir, method, matching_method, matching_params=[], unknown_painting_threshold=2,
+                    cache_segmented=False, cache_denoised=False):
+    
+    if cache_segmented:
+        print("Using cached segmented images.")
+        
+    if cache_denoised:
+        print("Using cached denoised images.")
     
     # REMOVE NOISE FROM QUERY IMAGES FOR SEGMENTATION
     # ======================================================================================
@@ -178,29 +185,30 @@ def get_predictions(query_dir, bbdd_dir, method, matching_method, matching_param
     # Create a new directory for denoised images
     denoised_for_segmentation_queries_dir = 'data/denoised_for_segmentation_queries_dir'
 
-    # Remove previous denoised images
-    if os.path.exists(denoised_for_segmentation_queries_dir):
-        shutil.rmtree(denoised_for_segmentation_queries_dir)
+    if not cache_segmented:
+        # Remove previous denoised images
+        if os.path.exists(denoised_for_segmentation_queries_dir):
+            shutil.rmtree(denoised_for_segmentation_queries_dir)
 
-    # Execute denoising method 1
-    create_denoised_dataset(
-        noisy_dataset_path = query_dir,
-        denoised_dataset_path = denoised_for_segmentation_queries_dir,
-        method='gaussian',
-        lowpass_params={'ksize': 3},
-        highpass=False
-    )
+        # Execute denoising method 1
+        create_denoised_dataset(
+            noisy_dataset_path = query_dir,
+            denoised_dataset_path = denoised_for_segmentation_queries_dir,
+            method='gaussian',
+            lowpass_params={'ksize': 3},
+            highpass=False
+        )
 
-    # Read denoised images
-    rgb_queries_denoised = []
-    for filename in os.listdir(denoised_for_segmentation_queries_dir):
-        if filename.endswith('.jpg'):
-            img_path = os.path.join(denoised_for_segmentation_queries_dir, filename)
-            img_rgb = cv2.imread(img_path)
-            if img_rgb is not None:
-                rgb_queries_denoised.append(img_rgb)
-            else:
-                print(f"Warning: Failed to read {img_path}")
+        # Read denoised images
+        rgb_queries_denoised = []
+        for filename in os.listdir(denoised_for_segmentation_queries_dir):
+            if filename.endswith('.jpg'):
+                img_path = os.path.join(denoised_for_segmentation_queries_dir, filename)
+                img_rgb = cv2.imread(img_path)
+                if img_rgb is not None:
+                    rgb_queries_denoised.append(img_rgb)
+                else:
+                    print(f"Warning: Failed to read {img_path}")
 
 
     # DETECT PAINTINGS IN QUERIES (SEGMENTATION)
@@ -209,73 +217,74 @@ def get_predictions(query_dir, bbdd_dir, method, matching_method, matching_param
     masks_queries_dir = "data/masks_queries_dir"
     cropped_queries_dir = "data/cropped_queries_dir"
 
-    # Remove previous masks and cropped images
-    if os.path.exists(masks_queries_dir):
-        shutil.rmtree(masks_queries_dir)
-    if os.path.exists(cropped_queries_dir):
-        shutil.rmtree(cropped_queries_dir)
+    if not cache_segmented:
+        # Remove previous masks and cropped images
+        if os.path.exists(masks_queries_dir):
+            shutil.rmtree(masks_queries_dir)
+        if os.path.exists(cropped_queries_dir):
+            shutil.rmtree(cropped_queries_dir)
 
-    # Create new directories for masks and cropped images
-    os.makedirs(masks_queries_dir, exist_ok=True)
-    os.makedirs(cropped_queries_dir, exist_ok=True)
+        # Create new directories for masks and cropped images
+        os.makedirs(masks_queries_dir, exist_ok=True)
+        os.makedirs(cropped_queries_dir, exist_ok=True)
 
-    masks = generate_masks(rgb_queries_denoised)
+        masks = generate_masks(rgb_queries_denoised)
 
-    # Read query images
-    rgb_queries = []
-    for filename in os.listdir(query_dir):
-        if filename.endswith('.jpg'):
-            img_path = os.path.join(query_dir, filename)
-            img_rgb = cv2.imread(img_path)
-            if img_rgb is not None:
-                rgb_queries.append(img_rgb)
-            else:
-                print(f"Warning: Failed to read {img_path}")
+        # Read query images
+        rgb_queries = []
+        for filename in os.listdir(query_dir):
+            if filename.endswith('.jpg'):
+                img_path = os.path.join(query_dir, filename)
+                img_rgb = cv2.imread(img_path)
+                if img_rgb is not None:
+                    rgb_queries.append(img_rgb)
+                else:
+                    print(f"Warning: Failed to read {img_path}")
 
-    paintings_per_image = []
-    image_counter = 0
+        paintings_per_image = []
+        image_counter = 0
 
-    for i, mask in enumerate(tqdm(masks, desc="Generating segmentation masks and cropping images")):
-        # Save the mask
-        mask_filename = f"{i:05d}.png"
-        masks_save_path = os.path.join(masks_queries_dir, mask_filename)
-        cv2.imwrite(masks_save_path, mask)
+        for i, mask in enumerate(tqdm(masks, desc="Generating segmentation masks and cropping images")):
+            # Save the mask
+            mask_filename = f"{i:05d}.png"
+            masks_save_path = os.path.join(masks_queries_dir, mask_filename)
+            cv2.imwrite(masks_save_path, mask)
 
-        # Detect connected components in the mask
-        num_labels, labels, stats, _ = cv2.connectedComponentsWithStats(mask, connectivity=8)
+            # Detect connected components in the mask
+            num_labels, labels, stats, _ = cv2.connectedComponentsWithStats(mask, connectivity=8)
 
-        # Count paintings (objects of interest) in the current image
-        painting_count = 0
+            # Count paintings (objects of interest) in the current image
+            painting_count = 0
 
-        # Collect all valid components (ignoring background label 0)
-        components = []
+            # Collect all valid components (ignoring background label 0)
+            components = []
 
-        for j in range(1, num_labels):
-            x, y, w, h, area = stats[j]
+            for j in range(1, num_labels):
+                x, y, w, h, area = stats[j]
 
-            components.append((x, y, w, h, area))  # Store component details
+                components.append((x, y, w, h, area))  # Store component details
 
-        # Sort components from leftmost to rightmost (higher x to lower x)
-        components_sorted = sorted(components, key=lambda c: c[0])
+            # Sort components from leftmost to rightmost (higher x to lower x)
+            components_sorted = sorted(components, key=lambda c: c[0])
 
-        # Iterate over sorted components and save them
-        for (x, y, w, h, area) in components_sorted:
-            # Extract and save the cropped region
-            cropped_result = rgb_queries[i][y:y+h, x:x+w]
-            cropped_filename = f"{image_counter:05d}.jpg"
-            cropped_save_path = os.path.join(cropped_queries_dir, cropped_filename)
-            cv2.imwrite(cropped_save_path, cropped_result)
+            # Iterate over sorted components and save them
+            for (x, y, w, h, area) in components_sorted:
+                # Extract and save the cropped region
+                cropped_result = rgb_queries[i][y:y+h, x:x+w]
+                cropped_filename = f"{image_counter:05d}.jpg"
+                cropped_save_path = os.path.join(cropped_queries_dir, cropped_filename)
+                cv2.imwrite(cropped_save_path, cropped_result)
 
-            # Increment the counter for unique naming
-            image_counter += 1
-            painting_count += 1
+                # Increment the counter for unique naming
+                image_counter += 1
+                painting_count += 1
 
-        # Add the number of paintings detected for this image to the list
-        paintings_per_image.append(painting_count)
+            # Add the number of paintings detected for this image to the list
+            paintings_per_image.append(painting_count)
 
-    # Save the list of frame counts per image in a single .pkl file
-    with open("data/paintings_per_image.pkl", "wb") as f:
-        pickle.dump(paintings_per_image, f)
+        # Save the list of frame counts per image in a single .pkl file
+        with open("data/paintings_per_image.pkl", "wb") as f:
+            pickle.dump(paintings_per_image, f)
 
 
     # REMOVE NOISE FROM QUERY IMAGES
@@ -287,21 +296,22 @@ def get_predictions(query_dir, bbdd_dir, method, matching_method, matching_param
 
     denoised_paintings_folder = 'data/denoised_paintings'
 
-    # Remove previous paintings
-    if os.path.exists(denoised_paintings_folder):
-        shutil.rmtree(denoised_paintings_folder)
+    if not cache_denoised:
+        # Remove previous paintings
+        if os.path.exists(denoised_paintings_folder):
+            shutil.rmtree(denoised_paintings_folder)
 
-    # Create new temporary directory for denoised images
-    os.makedirs(denoised_paintings_folder, exist_ok=True)
+        # Create new temporary directory for denoised images
+        os.makedirs(denoised_paintings_folder, exist_ok=True)
 
-    # Using denoising method 5
-    create_denoised_dataset(
-        noisy_dataset_path = cropped_queries_dir,
-        denoised_dataset_path = denoised_paintings_folder,
-        method='wavelet',
-        wavelet_params={'wavelet':'db1', 'mode':'soft', 'rescale_sigma':True},
-        highpass=False
-    )
+        # Using denoising method 5
+        create_denoised_dataset(
+            noisy_dataset_path = cropped_queries_dir,
+            denoised_dataset_path = denoised_paintings_folder,
+            method='wavelet',
+            wavelet_params={'wavelet':'db1', 'mode':'soft', 'rescale_sigma':True},
+            highpass=False
+        )
 
 
     # EXTRACT LOCAL FEATURES (KEYPOINT DESCRIPTORS) FROM
@@ -316,7 +326,6 @@ def get_predictions(query_dir, bbdd_dir, method, matching_method, matching_param
     query_key_des_list = get_key_des_multi_image(query_images, method)
     bbdd_key_des_list = get_key_des_multi_image(bbdd_images, method)
     
-    # return query_key_des_list, bbdd_key_des_list
     
     # GET PREDICTIONS USING MATCHING DESCRIPTORS
     # =======================================================================================
@@ -331,7 +340,7 @@ def get_predictions(query_dir, bbdd_dir, method, matching_method, matching_param
         num_matching_descriptors_list = []
         for bbdd_image in bbdd_key_des_list:
             
-            # There must be at least one descriptor
+            # There must be at least one descriptor in the bbdd image
             if str(bbdd_image['descriptors'])!="None":
                 _, num_matching_descriptors = get_num_matching_descriptors(query_image['descriptors'],
                                                                            bbdd_image['descriptors'],
@@ -365,5 +374,7 @@ results = get_predictions(query_dir, bbdd_dir,
                           method="ORB",
                           matching_method="BruteForce",
                           matching_params=[cv2.NORM_HAMMING, True],
-                          unknown_painting_threshold=None)
+                          unknown_painting_threshold=2,
+                          cache_segmented=True,
+                          cache_denoised=True)
 """
